--- app.yaml
@@ -1,17 +1,0 @@
-runtime: python37
-# env: flex
-# entrypoint: gunicorn -b :$PORT main:app
-
-# runtime_config:
-#   python_version: 3
-
-# # This sample incurs costs to run on the App Engine flexible environment.
-# # The settings below are to reduce costs during testing and are not appropriate
-# # for production use. For more information, see:
-# # https://cloud.google.com/appengine/docs/flexible/python/configuring-your-app-with-app-yaml
-# manual_scaling:
-#   instances: 1
-# resources:
-#   cpu: 1
-#   memory_gb: 0.5
-#   disk_size_gb: 10
