--- config.py
@@ -1,2 +1,0 @@
-URL = "http://127.0.0.1"
-NODE_PORT = "5100"
