--- friendly_nodes.config
@@ -1,1 +1,0 @@
-http://127.0.0.1:5100
