--- keys/TEST_checkpoints_private_keys
@@ -1,3 +1,0 @@
-008d1f80536cd8df3948f4a0d54565f0b65dc8a9f8969fb979a2f21ab1f21e05
-383bbf98b6b828276adf9cd724ce89814a156c3785e9076da4311cb8ac522cd8
-c952f3d3ae901e7cb7cb904cc96e91ff52be72e975fd966c8dec3315d20819f1
