--- keys/TEST_validators_private_keys
@@ -1,3 +1,0 @@
-2a9c2b803fa9805a77cdfb542f9cea821dcdc552a818d8ff1b3d8dcc3bf34f27
-014e5842941509eeb14415d1e90044a62c28fc84df419a98adf41867b93d9292
-6ce8c0f11e5fbe859ad860f2cb94779e4f1d13ac50d04e097df9247620b215b9
