--- keys/checkpoints_public_keys
@@ -1,3 +1,0 @@
-0404fb2416c38f8e0e4790973d6cfcae0bffd02db79f651ecba976f55e84406d49218d39cb1adee8a3a911ddfe0fae85491e990d48a8ce451224ab32143c8ac736
-0461ada792d87027998bde4508bddc325453a4c2d182c5f8298f11bfc39aa71d9dee96c233228fbf225b38c397fe2e6a3dc684ebef73939ec1948374844d827dd6
-044be5a56d65a3ee10910257fe25f5b2c00ea624c6ee50b81e789977401252fa7a7d272db909db3855d6f7119b0e32dcc2fffc73b8ebe1a175eb97993488a75b4f
