--- keys/validators_public_keys
@@ -1,3 +1,0 @@
-0441d5c43eaf96ad21bdfdfe2e88f57e3f079e683321358b03d5ffed7ac1ccbe956f9fb4a9a3f3c570121b69fe25c87e5dd6df87636e35d0d230e1d29201c3be0a
-042afd6ae0c48f8812600bd8cbc5a9fdf800c9fd9cd5d6ebb92f33e5939dc138b9412bf29dd1804f088946338203ded7577f0b33f6134a7660de5d2ccf1479af93
-04f169d4c29d623b35de538c2c2ad36c3ffb6157fb00e592b6bda2b8af68472d4196ca4fceaf7044190fd3ad9ac0e3a61160600f08642365708b1f39f82fd599fd
