+++ node_config/path_config.py
@@ -1,0 +1,2 @@
+CONFIG_PATH='node_config/'
+WRITE_PATH='node_tmp/'
