--- node_keys/checkpoint_private_key.wif
@@ -1,1 +1,0 @@
-5HpXfPm94cFqqZnQwZ5QnLrb4adRU9CrJXtxYWETxcnpWAPir61
