--- node_keys/validator_private_key.wif
@@ -1,1 +1,0 @@
-5J93zpinw6L4VMABJMUmJdDRUK8GVpkbncHXkTG69fEwf6YrtjN
