+++ node_tmp/0000.block
@@ -1,0 +1,1 @@
+{"timestamp": "1591199519.055152", "previous_hash": "0000000000000000000000000000000000000000000000000000000000000000", "transactions": ["00000001345890370001588410854.3013510025000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000cargo_info0404fb2416c38f8e0e4790973d6cfcae0bffd02db79f651ecba976f55e84406d49218d39cb1adee8a3a911ddfe0fae85491e990d48a8ce451224ab32143c8ac7368f575cdca1d5101ddf6ff105739e0f02c7997dce288a8ee512dcd76ce5ba89e53e1ddb64d7c5496aa6dacdca7a911ffdc90d5474c66c5363146af6caac3ee350"], "merkle_root": "9048482871627f0bd47137a78922f55d4097e0851cf0d219817aa7ec667dac52", "hash": "2f54ba0a08a25a70c8079e0022f7570cee96b62df3055d1b33b6d3826d0eb13d", "public_key": "0441d5c43eaf96ad21bdfdfe2e88f57e3f079e683321358b03d5ffed7ac1ccbe956f9fb4a9a3f3c570121b69fe25c87e5dd6df87636e35d0d230e1d29201c3be0a", "signed_hash": "db0d26be2a6cd79b329811c7aafcf34c31071e852361526f64a50ae2861ced75a87ae520f22999706422deb752bbf240495e2ab4797cc842a455a4c443c3feb4"}
