--- path_config.py
@@ -1,2 +1,0 @@
-PATH=''
-WRITE_PATH='/tmp/'
