--- public_address
@@ -1,1 +1,0 @@
-1FnK72hrb6SY4T5iGTdVDBrfijgBqbf1aH
